requirements.txt
gradio
torch
pandas
numpy
pytorch_lightning
tqdm
sckit-learn
matplotlib
Pillow
torchvision